Move all employee pages and components into his sub-component
Add mode button into employee pages
Move app class into html template
Fix key prop
