+++ onboarding/src/Components/Employee/NavbarEmployee.js
@@ -5,12 +5,12 @@ function NavbarEmployee() {
     return(
       <nav className="top-bar navbar navbar-expand-lg navbar-dark bg-dark">
         <div className="top-bar-brand"><a className="navbar-brand">Online Onboarding</a></div>
-        
+
         <div className="top-bar-list" id="navbarSupportedContent">
             <div className="top-bar-item top-bar-item-full">
                 <ul className="navbar-nav d-flex breadcrumb my-2 my-lg-0">
                     <li className="breadcrumb-item">
-                        <Link to="/employe_forms_list">Pulpit</Link>
+                        <Link to="/">Pulpit</Link>
                     </li>
                     <li className="breadcrumb-item active">
                         tytuł formularza
@@ -26,7 +26,7 @@ function NavbarEmployee() {
                         <button className="btn-account d-none d-md-flex" type="button" data-toggle="dropdown"><p><em>Witaj employee-name #login</em></p></button>
                         <div className="dropdown-menu">
                             <div className="dropdown-arrow ml-3"></div>
-                            <a className="dropdown-item" href={process.env.PUBLIC_URL+"/accounts/logout/"}><span className="dropdown-icon oi oi-account-logout"></span> Logout</a>
+                            <a className="dropdown-item" href={"/accounts/logout/"}><span className="dropdown-icon oi oi-account-logout"></span> Logout</a>
                             <div className="dropdown-divider"></div>
                         </div>
                     </div>
